Moved SILO io to its io subdirectory
